requirements.txt
Flask
openai
azure-cognitiveservices-speech
python-dotenv